requirements.txt
alembic==1.6.2
bcrypt==3.2.0
blinker==1.4
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cryptography==3.4.7
cssselect2==0.4.1
defusedxml==0.7.1
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-Dance==5.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WeasyPrint==0.6
Flask-WTF==0.14.3
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
MarkupSafe==2.0.0
oauthlib==3.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
pyOpenSSL==20.0.1
Pyphen==0.10.0
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.15
tinycss2==1.1.0
urllib3==1.26.4
URLObject==2.4.3
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.0
WTForms==2.3.3
